# api/openapi.yaml
openapi: 3.0.3
info:
  title: argus-agents User Management API
  version: 0.1.0
  description: |
    OAuth2/OIDC authentication + LLM provider management for argus-agents.

    ## Authentication

    All endpoints except `/auth/*` require a JWT Bearer token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Obtain a token via `GET /auth/callback` (dev mode) or the full OIDC flow via `/auth/login`.

    ## Dev Mode

    Set `DEV_MODE=true` in environment. Dev mode bypasses OIDC and uses the `X-Dev-User` header:

    ```http
    GET /auth/callback?code=dev&state=dev
    X-Dev-User: alice
    ```
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - Bearer: []
...
paths:
...
  /providers/test-config:
    post:
      operationId: test_provider_config
      summary: Test provider connectivity with raw config
      description: |
        Uses the submitted `api_key`, optional `base_url`, and optional top-level `model`
        to perform a connectivity probe without creating a persisted provider.
      tags: [providers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestConfigInput"
            example:
              api_key: sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
              base_url: https://api.openai.com/v1
              model: gpt-4o
      responses:
        "200":
          description: Connectivity test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /internal-providers:
    get:
      operationId: list_internal_providers
      summary: List all internal (system) providers
      description: |
        Returns all providers with `kind=internal`. The `config` field is never returned
        (configs are encrypted at rest).
      tags: [providers]
      responses:
        "200":
          description: List of internal providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderRead"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers:
    get:
      operationId: list_user_providers
      summary: List current user's provider associations
      description: |
        Returns all user-provider associations for the authenticated user, ordered by
        `created_at` descending (most recent first). Only includes `kind=user` providers.
        Internal providers are not included.
      tags: [providers]
      responses:
        "200":
          description: User's provider associations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProviderRead"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: create_user_provider
      summary: Create a new LLM provider for the current user
      description: |
        The `config` is JSON-serialized and encrypted with AES-256-GCM before storage.
        The encrypted value is never returned in any response.

        If this is the user's first provider, it is automatically set as the default.
      tags: [providers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCreate"
            example:
              name: My OpenAI
              config:
                api_key: sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
                base_url: https://api.openai.com
              models:
                - gpt-4o
                - gpt-4o-mini
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderRead"
              example:
                id: "b2c3d4e5-f6a7-8901-bcde-f23456789012"
                name: My OpenAI
                kind: user
                created_at: "2026-03-27T10:05:00Z"
                updated_at: "2026-03-27T10:05:00Z"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}:
    delete:
      operationId: delete_user_provider
      summary: Delete a user-provider association
      description: |
        Removes the association between the current user and the specified provider.
        The underlying `Provider` record is NOT deleted (it may be shared across users).

        If the deleted provider was the user's default, the most recently created
        remaining provider is automatically promoted to default.

        Attempting to delete an internal provider returns `403 Forbidden`.
        Attempting to delete another user's provider returns `404 Not Found`.
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
          description: Provider UUID
      responses:
        "204":
          description: Provider association deleted (no response body)
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Cannot delete internal provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider not found or not associated with user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}/default:
    put:
      operationId: set_default_provider
      summary: Set a provider as the user's default
      description: |
        Sets the specified provider as the user's default. All other `is_default` flags
        for this user are cleared.
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
          description: Provider UUID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefaultUpdate"
      responses:
        "200":
          description: Provider set as default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProviderRead"
              example:
                id: "b2c3d4e5-f6a7-8901-bcde-f23456789012"
                name: My OpenAI
                kind: user
                is_default: true
                created_at: "2026-03-27T10:05:00Z"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider not associated with user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}/models:
    get:
      operationId: list_provider_models
      summary: List all models under a provider
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Provider model list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderModelRead"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: add_provider_model
      summary: Add a model to a provider
      description: |
        Appends a model under the specified provider. If this is the first model for the
        provider, it becomes the default automatically.
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderModelCreate"
            example:
              name: gpt-4o-mini
      responses:
        "201":
          description: Provider model created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderModelRead"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate model name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}/models/{model_id}:
    delete:
      operationId: delete_provider_model
      summary: Delete a model from a provider
      description: |
        Deletes the target model. If the deleted model was the default and there are
        remaining models, the most recently created remaining model becomes the new default.
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: model_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Provider model deleted
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider or model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}/models/{model_id}/default:
    put:
      operationId: set_default_provider_model
      summary: Set the default model for a provider
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: model_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Provider model updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderModelRead"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider or model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /providers/{provider_id}/models/{model_id}/test:
    post:
      operationId: test_provider_model
      summary: Test a specific provider model
      description: |
        Uses the provider's persisted `api_key` and `base_url` together with the selected
        model name to perform a connectivity probe.
      tags: [providers]
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: model_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Connectivity test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Provider or model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/auth/callback`
  schemas:
...
    ProviderRead:
      type: object
      required: [id, name, kind, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-8901-bcde-f23456789012"
        name:
          type: string
          example: argus-internal-default
        kind:
          type: string
          enum: [user, internal]
          example: internal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProviderRead:
      type: object
      required: [id, name, kind, is_default, created_at, model_count]
      properties:
        id:
          type: string
          format: uuid
          description: Provider UUID
          example: "b2c3d4e5-f6a7-8901-bcde-f23456789012"
        name:
          type: string
          example: My OpenAI
        kind:
          type: string
          enum: [user, internal]
          example: user
        is_default:
          type: boolean
          description: Whether this is the user's default provider
          example: true
        created_at:
          type: string
          format: date-time
          description: When this association was created
        default_model_name:
          type: string
          nullable: true
          description: Default model name currently used for provider-level testing
          example: gpt-4o
        model_count:
          type: integer
          minimum: 0
          description: Number of configured models under this provider
          example: 2
    ProviderCreate:
      type: object
      required: [name, config]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: Display name for the provider
          example: My OpenAI
        config:
          $ref: "#/components/schemas/ProviderConfigInput"
        models:
          type: array
          description: Optional initial model names to create with the provider
          items:
            type: string
            minLength: 1
            maxLength: 128
          example: [gpt-4o, gpt-4o-mini]
    ProviderConfigInput:
      type: object
      required: [api_key]
      properties:
        api_key:
          type: string
          minLength: 1
          description: API key for the LLM provider
          example: sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        base_url:
          type: string
          nullable: true
          description: Optional custom base URL
          example: https://api.openai.com
    ProviderModelCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: Provider model name
          example: gpt-4o
    ProviderModelRead:
      type: object
      required: [id, name, is_default, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: gpt-4o
        is_default:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TestConfigInput:
      type: object
      required: [api_key]
      properties:
        api_key:
          type: string
          minLength: 1
          description: API key for the LLM provider
          example: sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        base_url:
          type: string
          nullable: true
          description: Optional custom base URL
          example: https://api.openai.com/v1
        model:
          type: string
          nullable: true
          description: Optional model name to test
          example: gpt-4o
    ProviderTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Connection successful
        latency_ms:
          type: integer
          nullable: true
          example: 235
    DefaultUpdate:
      type: object
      description: Empty body for PUT /providers/{id}/default
      default: {}
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Missing X-Dev-User header in dev mode